--- a/lib/internal/http2/core.js
+++ b/lib/internal/http2/core.js
@@ -2498,6 +2498,10 @@ function initializeOptions(options) {
   options.Http1ServerResponse = options.Http1ServerResponse ||
     http.ServerResponse;
 
+  options.Http2ServerRequest = options.Http2ServerRequest ||
+                                       Http2ServerRequest;
+  options.Http2ServerResponse = options.Http2ServerResponse ||
+                                        Http2ServerResponse;
   return options;
 }
 
@@ -2563,7 +2567,11 @@ class Http2Server extends NETServer {
 function setupCompat(ev) {
   if (ev === 'request') {
     this.removeListener('newListener', setupCompat);
-    this.on('stream', onServerStream);
+    this.on('stream', onServerStream.bind(
+      this,
+      this[kOptions].Http2ServerRequest,
+      this[kOptions].Http2ServerResponse)
+    );
   }
 }