@jasnell 2017-09-24
Definitely good with this approach in general but will do a review on Monday.