@mcollina 2017-09-25
@dougwilson is this something that you could make some use of in Express. Do you need it also for `require('http')`?

Express would be the main consumer of this feature, and if do not need it I'm 👎  as it increases the API surface area.