@hekike 2017-09-25
@mcollina `restify` could benefit from it as well. The current approach extends the Node Core's `IncomingMessage` and `ServerResponse` prototypes. See https://github.com/restify/node-restify/blob/master/lib/request.js#L19  cc @yunong