@apapirovski 2017-09-25
If this is going to be added to `http2`, it will need to be `allowHTTP1` aware (so accept IncomingMessage & ServerResponse) which means even larger API surface.