@apapirovski 2017-09-25
ping @Fishrock123 I saw you in https://github.com/expressjs/express/issues/2812 and you also did https://github.com/pillarjs/extend-proto so figured you might have some thoughts.