@hekike 2017-09-25
@apapirovski thanks for checking it. I fixed your suggestions, here is the latest benchmark output after the changes:

**By requests:**

`cat scatter.csv | Rscript benchmark/scatter.R --xaxis requests --category version --plot scatter-plot.png --log`

```
aggregating variable: streams

 requests        version     rate confidence.interval
     5000       original 6017.385           103.08612
     5000         params 6005.038            94.99691
     5000 setPrototypeOf 5899.892           116.80510
    10000       original 7066.274           130.80891
    10000         params 7133.153           118.89924
    10000 setPrototypeOf 7028.677           133.74257
    20000       original 7684.556           144.31313
    20000         params 8041.344           119.24608
    20000 setPrototypeOf 7791.633           111.70869
    50000       original 7754.759           116.79295
    50000         params 8187.066           123.17560
    50000 setPrototypeOf 7827.398           102.14100
```

X axis: total number of requests
Y axis: rate of operations

![scatter-plot-requests-3](https://user-images.githubusercontent.com/1764512/30807960-6681160c-a1fc-11e7-938c-6ae47f8a03bb.png)

**By streams:**

`cat scatter.csv | Rscript benchmark/scatter.R --xaxis streams --category version --plot scatter-plot.png --log`

```
aggregating variable: requests

 streams        version     rate confidence.interval
     100       original 7030.053            183.8673
     100         params 7270.266            232.4791
     100 setPrototypeOf 6951.486            232.5603
     500       original 7154.344            214.5401
     500         params 7466.283            210.6520
     500 setPrototypeOf 7216.214            203.2741
    1000       original 7274.672            187.8303
    1000         params 7414.865            200.5134
    1000 setPrototypeOf 7285.713            183.2550
    2500       original 7206.308            189.6652
    2500         params 7301.195            216.3267
    2500 setPrototypeOf 7090.802            213.5673
    5000       original 7002.380            216.8112
    5000         params 7312.006            192.2240
    5000 setPrototypeOf 7113.685            183.5417
   10000       original 7116.705            193.1916
   10000         params 7285.285            214.3297
   10000 setPrototypeOf 7163.500            184.8569
```

X axis: concurrency level
Y axis: rate of operations

![scatter-plot-streams-3](https://user-images.githubusercontent.com/1764512/30807961-6688297e-a1fc-11e7-97d1-7e00ef142cf1.png)