@evanlucas 2017-09-25
I would be +1 on doing this for http as well as we could use it for our internal http framework too