@mcollina 2017-09-25
Can we move the discussion on adding this for `require('http')`? I think we might get the best value there rather than in `require('http2')`. Then http2 will follow http.