@hekike 2017-09-25
@mcollina good idea. The only reason why I started with `http2` because I was working on restify's HTTP/2 support.