@dougwilson 2017-09-25
@mcollina at least with Express current design, it would not be able to use this. The main things I see are

1. This can only be set when constructing thr HTTP server instance, which is outside of Express since it sits at a higher level.
2. The prototype Express sets is app-specific. This means that if I have sub apps, the prototype will change as the request is entering / leaving the sub apps.