@mcollina 2017-09-25
@hekike would this means that this functionality is needed only by Restify?  I'm trying to understand who needs this functionality.