@yunong 2017-09-25
@mcollina This is something that is generally useful for most frameworks since it doesn’t require directly modifying the prototype of the `request` and `response` objects. Sure this will require the frameworks to pick up this new paradigm, but it is much better than directly modifying the prototype, which is what both restify and express does today. 

If this lands restify itself will be moving to using this approach, for both HTTP/1 and HTTP/2. This will also allow us to provide HTTP/2 support in restify quickly, as we have a PR lined up and ready to go that will use this approach.