@apapirovski 2017-09-25
@yunong how do you deal with `allowHTTP1: true` on `createSecureServer` in that situation? To me that seems to be the missing piece here.