@apapirovski 2017-09-25
Well, there needs to be a solution for http1 but also there needs to be a solution for passing through http1 req & res prototypes through http2 to http1 when a user creates a server with `allowHTTP1: true`. See https://nodejs.org/dist/latest-v8.x/docs/api/http2.html#http2_http2_createsecureserver_options_onrequesthandler and https://github.com/nodejs/node/blob/4f4d55d10e24d2ae31103527cc380f359f34173f/lib/internal/http2/core.js#L2258