@mcollina 2017-09-25
@hekike can you open up a PR for http? Then we will adjust this accordingly to support both http1 and http2.