@hekike 2017-09-28
@mcollina @apapirovski I extended this PR for HTTP/1 with the HTTP/2 fallback logic as well.
I also added some tests. Please review it.