@mcollina 2017-10-02
This is a substantial new feature, it is probably good if other @nodejs/tsc members review it.