@jasnell 2017-10-02
For http/1 I would agree, but this should be fine to add to the http/2 impl :-)