@apapirovski 2017-10-02
Right now this PR combines both http1 & http2 changes. But also the http2 version doesn't quite work without the http1 version given `allowHTTP1`. I believe @hekike would have to split out everything related to http1 to land this now?