@jasnell 2017-10-02
ah right... ok, forgot that they were combined. In that case, yeah, we should give @nodejs/tsc a couple days to review. If there are no objections by end of week (to give extra time due to the conference in vancouver this week) then we should be good.