@hekike 2017-10-03
@mcollina @jasnell @apapirovski I split them, here is the PR for HTTP/1:
https://github.com/nodejs/node/pull/15752