@hekike 2017-10-19
@apapirovski thanks, I fixed the outstanding issues.

Do you want to merge the "inheritance benchmark" commit as well?
It was more for demonstrating the solution. I'm not sure that it can be useful in the future.