@hekike 2017-10-19
@apapirovski I added the comment to the benchmark file in the case we want to keep it.

The job failed during initialization.