@apapirovski 2017-10-19
Thanks for the update!

CI is all good, failures are known issues that are being resolved.