@addaleax 2017-10-30
@nodejs/http2 Is there anything left to do for this to land?